requirements.txt
sqlalchemy
psycopg2-binary
alembic
python-dotenv
PyQt6